Add some quick notes on technical plans
A few of the engineers today noted that we're not all in sync about how we
want notice-and-comment to work and where we want it to go, from a technical
side. This is an attempt to sketch out a small component of that vision,
specifically around the types of data we use and where they come from. What of
this needs to be replaced? What needs to be expanded?