Add 0000_0000 content; remove unneeded intro
In preparation for the not-yet-published rule, this creates _empty_ CFR
changes and a set of mock meta data. Eventually both should be derivable from
the `notice` structure. In the meantime, this will allow the parsed preamble
to display.

This also removes the "intro" content to  the 2016_02749 proposal, as we parse
that now. We'll need to follow up with a data update when this is merged.